chore: filename -> FullDocumentPath (#1538)
