+++ e2e/testcases/v2-test-data/birth-declaration.ts
@@ -1,8 +1,6 @@
 import { v4 as uuidv4 } from 'uuid'
 import { GATEWAY_HOST } from '../../constants'
 import { faker } from '@faker-js/faker'
-import fs from 'fs'
-import path from 'path'
 import { getAllLocations, getLocationIdByName } from '../birth/helpers'
 import { createClient } from '@opencrvs/toolkit/api'
 import {
@@ -11,6 +9,7 @@ import {
   ActionUpdate,
   AddressType
 } from '@opencrvs/toolkit/events'
+import { getSignatureFile, uploadFile } from './utils'
 
 type InformantRelation = 'MOTHER' | 'BROTHER'
 
@@ -146,39 +145,6 @@ export type Declaration = Awaited<ReturnType<typeof getDeclaration>>
 export interface CreateDeclarationResponse {
   eventId: string
   declaration: Declaration
-}
-
-function getSignatureFile() {
-  const buffer = fs.readFileSync(path.join(__dirname, 'signature.png'))
-  return new File([buffer], `signature-${Date.now()}.png`, {
-    type: 'image/png'
-  })
-}
-
-async function uploadFile(file: File, token: string) {
-  const formData = new FormData()
-  const transactionId = uuidv4()
-  formData.append('file', file)
-  formData.append('transactionId', transactionId)
-
-  const url = new URL('/upload', GATEWAY_HOST)
-  const res = await fetch(url, {
-    method: 'POST',
-    headers: {
-      Authorization: `Bearer ${token}`
-    },
-    body: formData
-  })
-
-  if (!res.ok) {
-    throw new Error(`Failed to upload file: ${res.statusText}`)
-  }
-
-  return {
-    filename: `${transactionId}.png`,
-    originalFilename: file.name,
-    type: file.type
-  }
 }
 
 export async function createDeclaration(
